Convert another test to Rust.

+++ core/src/model/database.rs
@@ -403,7 +403,7 @@ pub trait Database {
     fn attribute_key_exists(
         &self,
         transaction: Option<Rc<RefCell<Transaction<Postgres>>>>,
-        form_id_in: i64,
+        form_id_in: i32,
         id_in: i64,
     ) -> Result<bool, anyhow::Error>;
     fn relation_type_key_exists(
